tests/wpt/web-platform-tests/css/css-ui/input-security-parsing.html

<!DOCTYPE html>
<title>Test parsing for input-security</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#input-security">
<link rel="author" title="Apple Inc">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
test_valid_value('input-security', 'initial');
test_valid_value('input-security', 'inherit');
test_valid_value('input-security', 'unset');
test_valid_value('input-security', 'revert');
test_valid_value('input-security', 'auto');
test_valid_value('input-security', 'none');

test_invalid_value('input-security', 'auto auto');
test_invalid_value('input-security', 'disc');
test_invalid_value('input-security', 'circle');
test_invalid_value('input-security', 'square');
</script>
