third_party/blink/web_tests/external/wpt/css/css-ui/input-security-parsing-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['input-security'] = "initial" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['input-security'] = "inherit" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['input-security'] = "unset" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['input-security'] = "revert" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['input-security'] = "auto" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[FAIL] e.style['input-security'] = "none" should set the property value
  assert_not_equals: property should be set got disallowed value ""
[PASS] e.style['input-security'] = "auto auto" should not set the property value
[PASS] e.style['input-security'] = "disc" should not set the property value
[PASS] e.style['input-security'] = "circle" should not set the property value
[PASS] e.style['input-security'] = "square" should not set the property value
Harness: the test ran to completion.
